Added UI and Player Death
Ammo UI and Death UI, with a retry button that restarts the scene and an exit button that closes the application. Also when the player hits 0 HP, the UI switches.

+++ Assets/Scripts/ChasingEnemy.cs
@@ -10,7 +10,7 @@ public class ChasingEnemy : MonoBehaviour
     [SerializeField] Vector3 RotationRate;
     [SerializeField] int LookRotationRange;
     [SerializeField] int DetectionRange;
-    [SerializeField] int AttackRange;
+    [SerializeField] float AttackRange;
     [SerializeField] int AttackDamage;
     private float RealRotationRangeMax;
     private float RealRotationRangeMin;
